[fix](case)fix manager show proc case unstable.
