+++ regression-test/suites/manager/test_manager_interface_1.groovy
@@ -496,128 +496,6 @@ DISTRIBUTED BY HASH(`k1`) BUCKETS 1"""))
     test_table_index()
 
 
-
-
-
-// show proc '/current_query_stmts'
-// show proc '/current_queries'
-// show processlist
-// kill query $query_id
-// SHOW PROC '/cluster_health/tablet_health' 
-    def test_proc = {
-
-        def futures = []
-        
-
-        futures.add( thread {
-
-            try{
-                sql """ select sleep(4.809); """
-            }catch(Exception e){
-                
-            }
-        })
-        futures.add( thread {
-            sleep(1000);
-            List<List<Object>> result = sql """ show proc '/current_query_stmts' """ 
-            logger.info("result = ${result}" )
-            def x = 0
-            def queryid = ""
-            logger.info("result = ${result}")
-
-            for( int i = 0;i<result.size();i++) {
-                if (result[i][7] != null && result[i][7].contains("sleep(4.809)") )//Statement
-                {
-                    x = 1
-                    queryid = result[i][0]
-                    logger.info("query ID = ${queryid}")
-                    assertTrue(result[i][0]!= null) //QueryId
-                    assertTrue(result[i][1]!= null) //ConnectionId
-                    assertTrue(result[i][2]!= null)//Catalog
-                    assertTrue(result[i][3]!= null)//Database
-                    assertTrue(result[i][4]!= null)//User
-                    assertTrue(result[i][5]!= null)//ExecTime
-                    assertTrue(result[i][5].isNumber())//ExecTime
-                    assertTrue(result[i][6]!= null)//SqlHash
-                }
-            }
-            assertTrue(x == 1)
-            
-            x = 0 
-            result = sql """  show proc '/current_queries' """
-            logger.info("result = ${result}")
-            for( int i = 0;i<result.size();i++) {
-                if (result[i][0] == queryid )//QueryId
-                {
-                    x = 1
-                    assertTrue(result[i][5]!= null)//ScanBytes
-                    assertTrue(result[i][6]!= null)//ProcessBytes
-                }
-            }
-            assertTrue(x == 1)
-
-            result = sql """  show processlist  """
-            logger.info("result = ${result}")
-            for( int i =0 ;i < result.size();i++ ){
-                assertTrue( result[i][2].toLowerCase() != "null"  )//User
-                assertTrue( result[i][3].toLowerCase() != "null"  )//Host
-                assertTrue( result[i][5].toLowerCase() != "null"  )//Catalog
-                assertTrue( result[i][6].toLowerCase() != "null"  )//Db
-                assertTrue( result[i][10].toLowerCase() != "null"  )//QueryId
-                if (result[i][10] == queryid) {
-                    x = 1
-                }
-            }
-
-            assertTrue(x == 1)
-            sql """ kill query "${queryid}" """
-            
-            x =  0
-            sleep(5000)
-
-            result = sql """  show proc '/current_queries' """
-            logger.info("result = ${result}")
-            for( int i = 0;i<result.size();i++) {
-                if (result[i][0] == queryid )//QueryId
-                {
-                    x = 1
-                }
-            }
-            assertTrue(x == 0)
-        })
-        futures.each { it.get() }
-
-
-        def tablet_num  = 0;
-        def healthy_num = 0;
-        def total_tablet_num  = 0;
-        def total_healthy_num = 0;
-        result = sql """  SHOW PROC '/cluster_health/tablet_health' """ 
-        logger.info("result = ${result}" )
-
-        for( int i =0 ;i < result.size();i++ ){
-            assertTrue(result[i][0].toLowerCase() != null ) // DbId
-            if (result[i][0].toLowerCase() == "total") {
-                total_tablet_num = result[i][2].toBigInteger();
-                total_healthy_num = result[i][3].toBigInteger();
-            }else {
-                tablet_num += result[i][2].toBigInteger();
-                healthy_num += result[i][3].toBigInteger();
-                
-            }
-            // assertTrue(result[i][2]()) // TabletNum
-            // assertTrue(result[i][3]()) // HealthyNum
-        }
-        assertTrue(total_healthy_num == healthy_num )
-        assertTrue(total_healthy_num == healthy_num )
-    
-
-
-    }
-    test_proc();
-
-
-
 //select a.*, b.*, c.NAME as WORKLOAD_GROUP_NAME from information_schema.active_queries a left join information_schema.backend_active_tasks b on a.QUERY_ID = b.QUERY_ID left join information_schema.workload_groups c on a.WORKLOAD_GROUP_ID = c.ID
     def  test_active_query = {
